Adding boolean flag to indicate broker id requirement.
