+++ cruise-control-client/cruisecontrolclient/client/Endpoint.py
@@ -36,6 +36,9 @@ class AbstractEndpoint(metaclass=ABCMeta):
     # when they may need to perform de-conflict checks on the targeted kafka cluster.
     can_execute_proposal: ClassVar[bool]
 
+    # Whether this endpoint requires the broker ids to be provided.
+    requires_broker_ids: ClassVar[bool]
+
     # An ordered collection of the known Parameter classes that can be instantiated for this Endpoint.
     available_Parameters: ClassVar[Tuple[CCParameter.AbstractParameter]]
 
@@ -160,6 +163,7 @@ class AddBrokerEndpoint(AbstractEndpoint):
     description = "Move partitions to the specified brokers, according to the specified goals"
     http_method = "POST"
     can_execute_proposal = True
+    requires_broker_ids = True
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.BrokerIdParameter,
@@ -185,16 +189,13 @@ class AddBrokerEndpoint(AbstractEndpoint):
         'kwargs': dict(aliases=['add_brokers', 'add-broker', 'add-brokers'], help=description)
     }
 
-    def __init__(self, broker_ids: Union[str, List[str]]):
-        AbstractEndpoint.__init__(self)
-        self.add_param("brokerid", broker_ids)
-
 
 class AdminEndpoint(AbstractEndpoint):
     name = "admin"
     description = "Used to change runtime configurations on the cruise-control server itself"
     http_method = "POST"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.ConcurrentLeaderMovementsParameter,
         CCParameter.ConcurrentPartitionMovementsPerBrokerParameter,
@@ -216,6 +217,7 @@ class BootstrapEndpoint(AbstractEndpoint):
     description = "Bootstrap the load monitor"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.ClearMetricsParameter,
         CCParameter.EndParameter,
@@ -233,6 +235,7 @@ class DemoteBrokerEndpoint(AbstractEndpoint):
     description = "Remove leadership and preferred leadership from the specified brokers"
     http_method = "POST"
     can_execute_proposal = True
+    requires_broker_ids = True
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.BrokerIdParameter,
@@ -253,10 +256,6 @@ class DemoteBrokerEndpoint(AbstractEndpoint):
         'kwargs': dict(aliases=['demote_brokers', 'demote-broker', 'demote-brokers'], help=description)
     }
 
-    def __init__(self, broker_ids: Union[str, List[str]]):
-        AbstractEndpoint.__init__(self)
-        self.add_param("brokerid", broker_ids)
-
 
 class FixOfflineReplicasEndpoint(AbstractEndpoint):
     # Warning, this Endpoint is only supported in kafka 1.1 and above
@@ -264,6 +263,7 @@ class FixOfflineReplicasEndpoint(AbstractEndpoint):
     description = "Fixes the offline replicas in the cluster (kafka 1.1+ only)"
     http_method = "POST"
     can_execute_proposal = True
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.ConcurrentLeaderMovementsParameter,
@@ -293,6 +293,7 @@ class KafkaClusterStateEndpoint(AbstractEndpoint):
     description = "Get under-replicated and offline partitions (and under MinISR partitions in kafka 2.0+)"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.TopicParameter,
         CCParameter.JSONParameter,
@@ -309,6 +310,7 @@ class LoadEndpoint(AbstractEndpoint):
     description = "Get the load on each kafka broker"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.JSONParameter,
@@ -319,15 +321,13 @@ class LoadEndpoint(AbstractEndpoint):
         'kwargs': dict(help=description)
     }
 
-    def __init__(self):
-        AbstractEndpoint.__init__(self)
-
 
 class PartitionLoadEndpoint(AbstractEndpoint):
     name = "partition_load"
     description = "Get the resource load for each partition"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.EndParameter,
@@ -351,6 +351,7 @@ class PauseSamplingEndpoint(AbstractEndpoint):
     description = "Pause metrics load sampling"
     http_method = "POST"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.JSONParameter,
         CCParameter.ReasonParameter,
@@ -367,6 +368,7 @@ class ProposalsEndpoint(AbstractEndpoint):
     description = "Get current proposals"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.DataFromParameter,
@@ -390,6 +392,7 @@ class RebalanceEndpoint(AbstractEndpoint):
     description = "Rebalance the partition distribution in the kafka cluster, according to the specified goals"
     http_method = "POST"
     can_execute_proposal = True
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.ConcurrentLeaderMovementsParameter,
@@ -415,15 +418,13 @@ class RebalanceEndpoint(AbstractEndpoint):
         'kwargs': dict(help=description)
     }
 
-    def __init__(self):
-        AbstractEndpoint.__init__(self)
-
 
 class RemoveBrokerEndpoint(AbstractEndpoint):
     name = "remove_broker"
     description = "Remove all partitions from the specified brokers, according to the specified goals"
     http_method = "POST"
     can_execute_proposal = True
+    requires_broker_ids = True
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.BrokerIdParameter,
@@ -460,6 +461,7 @@ class ResumeSamplingEndpoint(AbstractEndpoint):
     description = "Resume metrics load sampling"
     http_method = "POST"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = {
         CCParameter.JSONParameter,
         CCParameter.ReasonParameter,
@@ -476,6 +478,7 @@ class ReviewEndpoint(AbstractEndpoint):
     description = "Create, approve, or discard reviews"
     http_method = "POST"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.ApproveParameter,
         CCParameter.DiscardParameter,
@@ -493,6 +496,7 @@ class ReviewBoardEndpoint(AbstractEndpoint):
     description = "View already-created reviews"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.JSONParameter,
         CCParameter.ReviewIDsParameter
@@ -508,6 +512,7 @@ class RightsizeEndpoint(AbstractEndpoint):
     description = "Rightsize the broker or partition count"
     http_method = "POST"
     can_execute_proposal = True
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.JSONParameter,
         CCParameter.TopicParameter,
@@ -525,6 +530,7 @@ class StateEndpoint(AbstractEndpoint):
     description = "Get the state of cruise control"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.JSONParameter,
         CCParameter.SubstatesParameter,
@@ -548,6 +554,7 @@ class StopProposalExecutionEndpoint(AbstractEndpoint):
     description = "Stop the currently-executing proposal"
     http_method = "POST"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.ForceStopParameter,
         CCParameter.JSONParameter,
@@ -564,6 +571,7 @@ class TopicConfigurationEndpoint(AbstractEndpoint):
     description = "Update the configuration of the specified topics"
     http_method = "POST"
     can_execute_proposal = True
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.AllowCapacityEstimationParameter,
         CCParameter.ConcurrentLeaderMovementsParameter,
@@ -594,6 +602,7 @@ class TrainEndpoint(AbstractEndpoint):
     description = "Train the linear regression model"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.EndParameter,
         CCParameter.JSONParameter,
@@ -610,6 +619,7 @@ class UserTasksEndpoint(AbstractEndpoint):
     description = "Get the recent user tasks from cruise control"
     http_method = "GET"
     can_execute_proposal = False
+    requires_broker_ids = False
     available_Parameters = (
         CCParameter.ClientIdsParameter,
         CCParameter.EndpointsParameter,
